Created adding orderedMeals to order

+++ GastroApp/Models/Meal.cs
@@ -5,8 +5,8 @@
         public int Id { get; set; }
         public string Name { get; set; }
         public string? Description { get; set; }
-        public double Price { get; set; }
-        public double VATRate { get; set;}
+        public float Price { get; set; }
+        public float VATRate { get; set;}
         public int CategoryId { get; set; }
         public Category Category { get; set; }
         public List<Order> Orders { get; set; } = new List<Order>();
